requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
dateparser==1.2.0
decorator==5.1.1
dill==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
executing==2.0.1
fcm-django==2.1.0
firebase-admin==6.5.0
flower==2.0.1
fonttools==4.53.0
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
humanize==4.9.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
imageio==2.34.2
incremental==22.10.0
ipython==8.25.0
ipywidgets==8.1.3
jedi==0.19.1
joblib==1.4.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kombu==5.3.7
llvmlite==0.43.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
MetaTrader5==5.0.4288
msgpack==1.0.8
mypy-extensions==1.0.0
numba==0.60.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
plotly==5.22.0
priority==1.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.6
regex==2024.5.15
requests==2.32.3
rsa==4.9
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.14.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tenacity==8.4.2
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
vectorbt==0.26.1
vine==5.1.0
wcwidth==0.2.13
widgetsnbextension==4.0.11
zope.interface==6.4.post2
